# forester-backend/build/build/swagger.json
{
    "openapi": "3.0.0",
    "components": {
        "examples": {},
        "headers": {},
        "parameters": {},
        "requestBodies": {},
        "responses": {},
        "schemas": {
            "User": {
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "passwordConfirm": {
                        "type": "string"
                    }
                },
                "required": [
                    "firstname",
                    "lastname",
                    "email",
                    "password",
                    "passwordConfirm"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "HttpStatusCode": {
                "enum": [
                    100,
                    101,
                    102,
                    103,
                    200,
                    201,
                    202,
                    203,
                    204,
                    205,
                    206,
                    207,
                    208,
                    226,
                    300,
                    301,
                    302,
                    303,
                    304,
                    305,
                    306,
                    307,
                    308,
                    400,
                    401,
                    402,
                    403,
                    404,
                    405,
                    406,
                    407,
                    408,
                    409,
                    410,
                    411,
                    412,
                    413,
                    414,
                    415,
                    416,
                    417,
                    418,
                    421,
                    422,
                    423,
                    424,
                    425,
                    426,
                    428,
                    429,
                    431,
                    451,
                    500,
                    501,
                    502,
                    503,
                    504,
                    505,
                    506,
                    507,
                    508,
                    510,
                    511
                ],
                "type": "number"
            },
            "ForesterError": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "stack": {
                        "type": "string"
                    },
                    "httpStatusCode": {
                        "$ref": "#/components/schemas/HttpStatusCode"
                    },
                    "isOperational": {
                        "type": "boolean",
                        "default": true
                    }
                },
                "required": [
                    "name",
                    "message",
                    "httpStatusCode"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "Job": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "areaPolygon": {
                        "items": {
                            "items": {
                                "items": {
                                    "type": "number",
                                    "format": "double"
                                },
                                "type": "array"
                            },
                            "type": "array"
                        },
                        "type": "array"
                    },
                    "email": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastCompletedRun": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "uuid",
                    "areaPolygon",
                    "email",
                    "status"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "JobGetDto": {
                "$ref": "#/components/schemas/Job"
            },
            "Pick_Job.areaPolygon-or-email_": {
                "properties": {
                    "areaPolygon": {
                        "items": {
                            "items": {
                                "items": {
                                    "type": "number",
                                    "format": "double"
                                },
                                "type": "array"
                            },
                            "type": "array"
                        },
                        "type": "array"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "required": [
                    "areaPolygon",
                    "email"
                ],
                "type": "object",
                "description": "From T, pick a set of properties whose keys are in the union K"
            },
            "JobCreateDto": {
                "$ref": "#/components/schemas/Pick_Job.areaPolygon-or-email_"
            },
            "JobUpdateDto": {
                "$ref": "#/components/schemas/Job"
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "forester-backend",
        "version": "1.0.0",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "George-Alexandru Tudurean ",
            "email": "atudurean@dvloper.io"
        }
    },
    "paths": {
        "/users": {
            "get": {
                "operationId": "GetUsers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User Controller Routes"
                ],
                "security": [],
                "parameters": []
            },
            "post": {
                "operationId": "CreateUser",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "User already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User Controller Routes"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                }
            }
        },
        "/users/{uuid}": {
            "get": {
                "operationId": "GetUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    ],
                                    "nullable": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User Controller Routes"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateUser",
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User Controller Routes"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteUser",
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "User Controller Routes"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/job": {
            "post": {
                "operationId": "CreateJob",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobGetDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Job Controller Routes"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobCreateDto"
                            }
                        }
                    }
                }
            },
            "get": {
                "operationId": "GetJobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/JobGetDto"
                                    },
                                    "type": "array"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Job Controller Routes"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/job/{jobId}": {
            "get": {
                "operationId": "GetJobById",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobGetDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Job Controller Routes"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "jobId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateJob",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobGetDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Job Controller Routes"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "query",
                        "name": "jobId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobUpdateDto"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteJob",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Job Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Job Controller Routes"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "query",
                        "name": "jobId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/job/{jobId}/publish": {
            "post": {
                "operationId": "PublishResults",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Job Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForesterError"
                                },
                                "examples": {
                                    "Example 1": {}
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Job Controller Routes"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "query",
                        "name": "jobId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}